Update kill.gs
added the kill command into the code. Now it will check for metaxploit.so and install it if its not present, start the rshell process, and finally kill whatever the user was trying to kill. 

The purpose, is if the player kills existing rshell, the kill command will just start a new one when to do it. this script will be used in tandem with the poisoned ps